+++ packages/sdk/test/ci/util.test.ts
@@ -1,5 +1,6 @@
-import { stringToInt, wasmToSchema } from '../../src/util.js';
+import { stringToInt } from '../../src/util.js';
 import { readFileSync } from 'fs';
+import { getEmbeddedModuleSchema } from '../../src/index.js';
 
 test('stringToInt transforms chosen field, but not others', () => {
     const keysToTransform = ['a'];
@@ -39,7 +40,12 @@ test('Embedded schema is the same as a seperate schema file', () => {
     const seperateSchema = readFileSync(
         'test/ci/resources/icecream-schema.bin'
     );
-    const embeddedSchema = wasmToSchema(wasmModule);
+    const embeddedSchema = getEmbeddedModuleSchema({
+        source: wasmModule,
+        version: 1,
+    });
 
-    expect(new Uint8Array(seperateSchema)).toEqual(embeddedSchema);
+    expect(new Uint8Array(seperateSchema)).toEqual(
+        new Uint8Array(embeddedSchema!.buffer)
+    );
 });